+++ NEWS.md
@@ -2,6 +2,7 @@
 # v0.1.1
 * Allow selection of enhanced profiles for plotting
 * Add argument `orientation` and enable landscape mode for plotting
+* Add new function `app_LESLIE()` to run the LESLIE shiny app
 
 # v0.1.0